// apps/web/src/types/metadata.ts
// Generated by https://quicktype.io

export interface Meta {
  data: MetaData;
  meta: MetaClass;
}

export interface MetaData {
  id: number;
  attributes: PurpleAttributes;
}

export interface PurpleAttributes {
  siteTitle: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
  links: Link[];
  social: Social[];
  logo: Favicon;
  favicon: Favicon;
}

export interface Favicon {
  data: FaviconData | null;
}

export interface FaviconData {
  id: number;
  attributes: FluffyAttributes;
}

export interface FluffyAttributes {
  name: string;
  alternativeText: string;
  caption: string;
  width: number;
  height: number;
  formats: Formats;
  hash: string;
  ext: string;
  mime: string;
  size: number;
  url: string;
  previewUrl: null;
  provider: string;
  provider_metadata: ProviderMetadata;
  createdAt: string;
  updatedAt: string;
}

export interface Formats {
  large: Large;
  small: Large;
  medium: Large;
  thumbnail: Large;
}

export interface Large {
  ext: string;
  url: string;
  hash: string;
  mime: string;
  name: string;
  path: null;
  size: number;
  width: number;
  height: number;
  provider_metadata: ProviderMetadata;
}

export interface ProviderMetadata {
  public_id: string;
  resource_type: string;
}

export interface Link {
  link: string;
  label: string;
}

export interface Social {
  url: string;
  icon: string;
}

export interface MetaClass {}